+++ pysradb/filter_attrs.py
@@ -3,6 +3,7 @@ import re
 import pandas as pd
 import numpy as np
 
+
 def _get_sample_attr_keys(sample_attribute):
     if sample_attribute is None:
         return None, None